pyproject.toml
[project]
name = "Argos"
version = "0.1.0"
description = "Argos – Detecção facial com OpenCV"
requires-python = ">=3.12"
dependencies = [
    "opencv-contrib-python>=4.10",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "numpy>=2.0",
    "loguru>=0.7.0",
    "ultralytics>=8.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "sentence-transformers>=2.0",
    "chromadb>=0.5.0",
    "google-genai>=1.40.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = "."
